requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Cython==3.0.12
exceptiongroup==1.2.2
Faker==37.1.0
Flask==3.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kafka-python==2.1.4
MarkupSafe==3.0.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-surprise==1.1.4
scipy==1.15.2
six==1.17.0
tomli==2.2.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
